requirements.txt
matplotlib==3.8.0
nltk==3.8.1
numpy==1.26.4
Pillow==10.3.0
streamlit==1.33.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.65.0